services/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "myhr-ai-api"
version = "0.1.0"
description = "FastAPI backend for the myhr-ai HR policy assistant"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "python-multipart>=0.0.9,<1.0.0",
  "openai>=1.40.0,<2.0.0",
  "azure-search-documents>=11.6.0b1,<12.0.0",
  "pypdf>=4.3.1,<5.0.0",
  "httpx>=0.27.0,<1.0.0",
  "orjson>=3.10.0,<4.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.23.0,<1.0.0",
  "mypy>=1.11.0,<2.0.0",
  "ruff>=0.6.0,<1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
